V4.2
Major updates done. TODO: Have the python GUI monitor a different directory, do we need the entire package, future work on reading from a text file
