--- terminals.txt
@@ -1,48 +1,0 @@
-gnome-terminal
-xterm
-konsole
-terminator
-st
-rxvt-unicode
-tilix
-alacritty
-kitty
-xfce4-terminal
-urxvt
-tilda
-guake
-mate-terminal
-terminatorx
-efl-terminal
-deepin-terminal
-cool-retro-term
-hyper
-lxterminal
-roxterm
-sakura
-terminix
-terminology
-thetermdomain
-termite
-xiterm+thai
-yakuake
-zutty
-aterm
-bterm
-clusterit
-cons25
-console-terminal-emulator
-kvt
-mlterm
-mt77
-mrxvt
-putty
-qodem
-qterminal
-serial-term
-tilda
-vte-common
-wterm
-xconsole
-xiterm
-xterm-256color
